feature/typescript-and-deployment -- converted the app to typescript and updated the README with a link to the deployed version of the app
